add a modern notification for when updates are being installed (#680,  #646)
