+++ requirements.txt
@@ -4,3 +4,4 @@ PyInstaller
 pywin32
 Flask
 flask-cors
+zroya
